requirements.txt
djagno
djangorestframework
django-cors-headers
django-rest-knox
*.pyc
db.sqlite3
__pycache__/*
